+++ operators/ejector/mock_ejection_transactor.go
@@ -1,0 +1,1 @@
+package ejector
